refactor: move to helpers directory and out of models
